Fix entity observaciones to get comment type

+++ src/main/java/cl/desagen/chilquinta/controllers/ObservacionNormaController.java
@@ -35,7 +35,7 @@ public class ObservacionNormaController {
     }
 
     @GetMapping(value = "getByIdNorma/{id}", produces = APPLICATION_JSON_UTF8_VALUE)
-    public Iterable<ObservacionNormaEntity> findAllByNormaId(@PathVariable Long id) {
+    public Iterable<ObservacionNormaEntity> findAllByNormaId(@PathVariable Integer id) {
         return observacionnormaService.findAllByNormaId(id);
     }
